Let `kotlin-dsl` plugin enable SAM conversions for Kotlin functions
