+++ subprojects/provider/src/test/kotlin/org/gradle/kotlin/dsl/integration/KotlinBuildScriptIntegrationTest.kt
@@ -1,6 +1,7 @@
 package org.gradle.kotlin.dsl.integration
 
 import org.gradle.kotlin.dsl.fixtures.AbstractIntegrationTest
+import org.gradle.kotlin.dsl.fixtures.containsMultiLineString
 
 import org.hamcrest.CoreMatchers.containsString
 
@@ -53,26 +54,7 @@ class KotlinBuildScriptIntegrationTest : AbstractIntegrationTest() {
     @Test
     fun `accepts lambda as SAM argument to Kotlin function`() {
 
-        withDefaultSettingsIn("buildSrc")
-
-        withBuildScriptIn("buildSrc", """
-
-            import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
-
-            plugins {
-                `kotlin-dsl`
-            }
-
-            tasks.withType<KotlinCompile> {
-                kotlinOptions {
-                    freeCompilerArgs += listOf(
-                        "-XXLanguage:+NewInference",
-                        "-XXLanguage:+SamConversionForKotlinFunctions"
-                    )
-                }
-            }
-
-        """)
+        withKotlinBuildSrc()
 
         withFile("buildSrc/src/main/kotlin/my.kt", """
             package my
@@ -81,19 +63,7 @@ class KotlinBuildScriptIntegrationTest : AbstractIntegrationTest() {
 
             fun <T> create(name: String, factory: org.gradle.api.NamedDomainObjectFactory<T>): T = factory.create(name)
 
-            fun <T : Any> createK(type: kotlin.reflect.KClass<T>, factory: org.gradle.api.NamedDomainObjectFactory<T>): T = factory.create(type.simpleName!!)
-
-            fun test() {
-
-                // Implicit SAM conversion in regular source
-                println(createK(String::class) { it.toUpperCase() })
-                println(create("FOO") { it.toLowerCase() })
-
-                // Implicit SAM with receiver conversion in regular source
-                applyActionTo("BAR") {
-                    println(toLowerCase())
-                }
-            }
+            fun <T : Any> create(type: kotlin.reflect.KClass<T>, factory: org.gradle.api.NamedDomainObjectFactory<T>): T = factory.create(type.simpleName!!)
         """)
 
         withBuildScript("""
@@ -105,22 +75,29 @@ class KotlinBuildScriptIntegrationTest : AbstractIntegrationTest() {
                     // Explicit SAM conversion
                     println(create("foo", NamedDomainObjectFactory<String> { it.toUpperCase() }))
                     // Explicit SAM conversion with generic type argument inference
-                    println(create<String>("foo", NamedDomainObjectFactory { it.toUpperCase() }))
+                    println(create<String>("bar", NamedDomainObjectFactory { it.toUpperCase() }))
                     // Implicit SAM conversion
-                    println(create<String>("foo") { it.toUpperCase() })
-                    println(createK(String::class) { it.toUpperCase() })
-                    println(createK(String::class, { name: String -> name.toUpperCase() }))
+                    println(create<String>("baz") { it.toUpperCase() })
+                    println(create(String::class) { it.toUpperCase() })
+                    println(create(String::class, { name: String -> name.toUpperCase() }))
                     // Implicit SAM with receiver conversion
                     applyActionTo("action") {
                         println(toUpperCase())
                     }
-                    test()
                 }
             }
          """.replaceIndent())
 
         assertThat(
-            build("test", "-q").output.also(::println),
-            containsString("ACTION"))
+            build("test", "-q").output,
+            containsMultiLineString("""
+                FOO
+                BAR
+                BAZ
+                STRING
+                STRING
+                ACTION
+            """)
+        )
     }
 }
